# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: MobiCare Pharmacy Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pharmacy/orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for this pharmacy
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders with items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
  /patient/orders:
    get:
      operationId: patientListOrders
      tags: [patient]
      summary: List the authenticated patient's own orders
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Patient orders with items, pharmacy, courier, prescription status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientOrder"
    post:
      operationId: patientCreateOrder
      tags: [patient]
      summary: Place an order (delivery or collection)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientOrderInput"
      responses:
        "201":
          description: Created order (awaiting payment)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientOrder"
        "409":
          description: Stock, tier, or availability conflict
  /patient/orders/{id}:
    get:
      operationId: patientGetOrder
      tags: [patient]
      summary: Get one of the patient's own orders
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order with items, pharmacy, courier, prescription status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientOrder"
...
  /hq/orders:
    get:
      operationId: listHqOrders
      tags: [hq]
      summary: All orders across all pharmacies
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HqOrder"
  /hq/orders/dispatch:
    get:
      operationId: listDispatchOrders
      tags: [hq]
      summary: Delivery orders in the dispatch pipeline
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Dispatch orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HqOrder"
  /hq/orders/{id}/assign-courier:
    post:
      operationId: assignCourier
      tags: [hq]
      summary: Assign a courier to a ready delivery order
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignCourierInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HqOrder"
        "409":
          description: Invalid state
  /hq/orders/{id}/courier-status:
    patch:
      operationId: updateCourierStatus
      tags: [hq]
      summary: Advance courier status (delivering / delivered)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HqOrder"
        "409":
          description: Invalid transition
  /hq/orders/{id}/cash-collected:
    post:
      operationId: markCashCollected
      tags: [hq]
      summary: Reconcile cash-on-delivery for a delivered order
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HqOrder"
        "409":
          description: Invalid state
...
  /hq/flags:
    get:
      operationId: listFlags
      tags: [hq]
      summary: Fraud/anomaly flags queue
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, reviewed]
      responses:
        "200":
          description: Flags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HqFlag"
  /hq/flags/{id}/review:
    post:
      operationId: reviewFlag
      tags: [hq]
      summary: Mark a flag reviewed with a note
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlagReviewInput"
      responses:
        "200":
          description: Reviewed flag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HqFlag"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderItem:
      type: object
      required: [id, orderId, drugId, drugName, quantity, unitPriceLeones]
      properties:
        id:
          type: string

        orderId:
          type: string

        drugId:
          type: string
        inventoryId:
          type: ["string", "null"]

        drugName:
          type: string
        quantity:
          type: number
        unitPriceLeones:
          type: number
        prescriptionId:
          type: ["string", "null"]
    Order:
      type: object
      required:
        [
          id,
          pharmacyId,
          patientName,
          patientPhone,
          status,
          fulfillmentType,
          idChecked,
          totalLeones,
          createdAt,
          updatedAt,
          items,
        ]
      properties:
        id:
          type: string

        pharmacyId:
          type: string

        patientName:
          type: string
        patientPhone:
          type: string
        status:
          type: string
          enum:
            [
              awaiting_payment,
              paid,
              confirmed,
              packaging,
              ready,
              assigned,
              picked_up,
              delivering,
              delivered,
              collected,
              cancelled,
            ]
        fulfillmentType:
          type: string
          enum: [delivery, collection]
        idChecked:
          type: boolean
        totalLeones:
          type: number
        prescriptionId:
          type: ["string", "null"]

        createdAt:
          type: string

        updatedAt:
          type: string

        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    # ── Patient schemas ────────────────────────────────────────────────────
...
    PatientOrderItemInput:
      type: object
      required: [inventoryId, quantity]
      properties:
        inventoryId:
          type: string
        quantity:
          type: number
          minimum: 1
    PatientOrderInput:
      type: object
      required: [pharmacyId, fulfillmentType, items]
      properties:
        pharmacyId:
          type: string
        fulfillmentType:
          type: string
          enum: [delivery, collection]
        deliveryAddress:
          type: string
        prescriptionImageKey:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PatientOrderItemInput"
    OrderPharmacyInfo:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    OrderCourierInfo:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
    OrderPrescriptionInfo:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectReason:
          type: ["string", "null"]
    PatientOrder:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            deliveryAddress:
              type: ["string", "null"]
            paymentMethod:
              type: string
            pharmacy:
              oneOf:
                - $ref: "#/components/schemas/OrderPharmacyInfo"
                - type: "null"
            courier:
              oneOf:
                - $ref: "#/components/schemas/OrderCourierInfo"
                - type: "null"
            prescription:
              oneOf:
                - $ref: "#/components/schemas/OrderPrescriptionInfo"
                - type: "null"
...
    HqCourierSummary:
      type: object
      required: [id, name, phone]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
    HqOrder:
      type: object
      required:
        [
          id,
          pharmacyId,
          patientName,
          patientPhone,
          status,
          fulfillmentType,
          idChecked,
          totalLeones,
          cashCollected,
          paymentMethod,
          createdAt,
          updatedAt,
          items,
        ]
      properties:
        id:
          type: string
        pharmacyId:
          type: string
        pharmacyName:
          type: ["string", "null"]
        patientName:
          type: string
        patientPhone:
          type: string
        status:
          type: string
          enum:
            [
              awaiting_payment,
              paid,
              confirmed,
              packaging,
              ready,
              assigned,
              picked_up,
              delivering,
              delivered,
              collected,
              cancelled,
            ]
        fulfillmentType:
          type: string
          enum: [delivery, collection]
        idChecked:
          type: boolean
        totalLeones:
          type: number
        prescriptionId:
          type: ["string", "null"]
        courierId:
          type: ["string", "null"]
        courier:
          oneOf:
            - $ref: "#/components/schemas/HqCourierSummary"
            - type: "null"
        cashCollected:
          type: boolean
        cashCollectedAt:
          type: ["string", "null"]
        paymentMethod:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    AssignCourierInput:
      type: object
      required: [courierId]
      properties:
        courierId:
          type: string
    CourierStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [delivering, delivered]
...
    HqFlag:
      type: object
      required: [id, orderId, type, reason, status, createdAt]
      properties:
        id:
          type: string
        orderId:
          type: string
        type:
          type: string
          enum: [velocity, duplicate, payment_anomaly]
        reason:
          type: string
        details: {}
        status:
          type: string
          enum: [open, reviewed]
        reviewNote:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        orderStatus:
          type: ["string", "null"]
        orderTotalLeones:
          type: ["number", "null"]
        patientName:
          type: ["string", "null"]
        patientPhone:
          type: ["string", "null"]
        pharmacyName:
          type: ["string", "null"]
    FlagReviewInput:
      type: object
      required: [note]
      properties:
        note:
          type: string
          minLength: 1
...